+++ extension/test/unit/user/address/add-test.js
@@ -5,7 +5,8 @@ describe('user / address / add', () => {
   const context = {
     meta: {
       userId: null
-    }
+    },
+    log: { debug: () => {} }
   }
 
   it('should throw an unauthorized error if no user is logged in', async () => {