+++ extension/user/address/add.js
@@ -1,5 +1,5 @@
 const UnauthorizedError = require('../../models/Errors/UnauthorizedError')
-const ApiFactory = require('../../lib/shopify.api.factory')
+const ApiFactory = require('../../lib/ShopifyApiFactory')
 const { mapCustomAttributes } = require('../../lib/mapper')
 
 /**
@@ -8,6 +8,7 @@ const { mapCustomAttributes } = require('../../lib/mapper')
  */
 module.exports = async (context, input) => {
   if (!context.meta.userId) {
+    context.log.debug('No user ID set in meta data')
     throw new UnauthorizedError('User is not logged in.')
   }
 
@@ -25,7 +26,7 @@ module.exports = async (context, input) => {
 
   const tokenManager = ApiFactory.buildShopifyApiTokenManager(context)
   const storefrontApi = ApiFactory.buildStorefrontApi(context, tokenManager)
-  const customerAccessToken = await tokenManager.getCustomerAccessToken()
+  const customerAccessToken = await tokenManager.getStorefrontApiCustomerAccessToken()
 
   return storefrontApi.customerAddressCreate(customerAccessToken.accessToken, newAddress)
 }