+++ extension/user/get.js
@@ -1,4 +1,4 @@
-const ApiFactory = require('../lib/shopify.api.factory')
+const ApiFactory = require('../lib/ShopifyApiFactory')
 const UnauthorizedError = require('../models/Errors/UnauthorizedError')
 const ShopgateCustomer = require('../models/user/ShopgateCustomer')
 
@@ -7,8 +7,8 @@ const ShopgateCustomer = require('../models/user/ShopgateCustomer')
  * @return {ShopgateCustomer}
  */
 module.exports = async function (context) {
-  // Check if there is a userId within the context.meta-data, if not the user is not logged
   if (!context.meta.userId) {
+    context.log.debug('No user ID set in meta data')
     throw new UnauthorizedError('Unauthorized user')
   }
 
@@ -18,20 +18,53 @@ module.exports = async function (context) {
     return userData.user
   }
 
-  let customerData = { id: null, mail: null }
-  try {
-    const tokenManager = ApiFactory.buildShopifyApiTokenManager(context)
-    const storefrontApi = ApiFactory.buildStorefrontApi(context, tokenManager)
-    const customerAccessToken = await tokenManager.getCustomerAccessToken()
+  const tokenManager = ApiFactory.buildShopifyApiTokenManager(context)
 
-    customerData = ShopgateCustomer.fromShopifyCustomer(await storefrontApi.getCustomerByAccessToken(customerAccessToken.accessToken))
+  let customerData
+  try {
+    const customerAccountApiAccessToken = await tokenManager.getCustomerAccountApiAccessToken()
+    customerData = customerAccountApiAccessToken
+      ? await _getCustomerFromCustomerAccountApi(context, customerAccountApiAccessToken)
+      : await _getCustomerFromStorefrontApi(context, tokenManager)
   } catch (err) {
-    return customerData
+    context.log.error({ errorMessage: err.message, code: err.code, statusCode: err.statusCode }, 'Error getting customer data')
+    throw err
   }
+
+  if (!customerData) return { id: null, mail: null }
+
   await context.storage.user.set('userData', {
-    ttl: (new Date()).getTime() + context.config.userDataCacheTtl, // cache for N microseconds
+    ttl: (new Date()).getTime() + context.config.userDataCacheTtl,
     user: customerData
   })
 
   return customerData
+}
+
+async function _getCustomerFromStorefrontApi (context, tokenManager) {
+  try {
+    const storefrontApi = ApiFactory.buildStorefrontApi(context, tokenManager)
+    const customerAccessToken = await tokenManager.getStorefrontApiCustomerAccessToken()
+
+    return ShopgateCustomer.fromShopifyStorefrontApiCustomer(await storefrontApi.getCustomerByAccessToken(customerAccessToken.accessToken))
+  } catch (err) {
+    context.log.error({ errorMessage: err.message }, 'Error getting customer data from Storefront API.')
+  }
+}
+
+async function _getCustomerFromCustomerAccountApi (context, customerAccountApiAccessToken) {
+  try {
+    const customerAccountApi = ApiFactory.buildCustomerAccountApi(context)
+    const customerDataResult = await customerAccountApi.getCustomer(customerAccountApiAccessToken.accessToken)
+
+    const customerData = ((customerDataResult || {}).data || {}).customer
+    if (!customerData) {
+      context.log.error({ customerDataResult: JSON.stringify(customerDataResult) }, 'Error getting customer data from Customer Account API.')
+      return
+    }
+
+    return ShopgateCustomer.fromShopifyCustomerAccountApiCustomer(customerData)
+  } catch (err) {
+    context.log.error({ errorMessage: err.message, code: err.code, statusCode: err.statusCode }, 'Error getting customer data from Customer Account API.')
+  }
 }