+++ extension/user/password/update.js
@@ -1,4 +1,4 @@
-const ApiFactory = require('../../lib/shopify.api.factory')
+const ApiFactory = require('../../lib/ShopifyApiFactory')
 const FieldValidationError = require('../../models/Errors/FieldValidationError')
 const UnauthorizedError = require('../../models/Errors/UnauthorizedError')
 
@@ -15,6 +15,7 @@ const UnauthorizedError = require('../../models/Errors/UnauthorizedError')
  */
 module.exports = async (context, input) => {
   if (!context.meta.userId) {
+    context.log.debug('No user ID set in meta data')
     throw new UnauthorizedError()
   }
 
@@ -26,12 +27,12 @@ module.exports = async (context, input) => {
 
   const tokenManager = ApiFactory.buildShopifyApiTokenManager(context)
   const storefrontApi = ApiFactory.buildStorefrontApi(context, tokenManager)
-  const customerAccessToken = await tokenManager.getCustomerAccessToken()
+  const customerAccessToken = await tokenManager.getStorefrontApiCustomerAccessToken()
 
   const options = { password: input.password }
 
   const result = await storefrontApi.updateCustomerByAccessToken(customerAccessToken.accessToken, options)
-  await tokenManager.setCustomerAccessToken(result.customerAccessToken)
+  await tokenManager.setStorefrontApiCustomerAccessToken(result.customerAccessToken)
 
   return result
 }